+++ platform/src/sdk/models/shared/connectionstatuschangedwebhookdata.ts
@@ -29,7 +29,7 @@ export class ConnectionStatusChangedWebhookData extends SpeakeasyBase {
     oldStatus?: DataConnectionStatus;
 
     /**
-     * A unique 4-letter key to represent a platform in each integration. View [accounting](https://docs.codat.io/integrations/accounting/accounting-platform-keys), [banking](https://docs.codat.io/integrations/banking/banking-platform-keys), and [commerce](https://docs.codat.io/integrations/commerce/commerce-platform-keys) platform keys.
+     * A unique 4-letter key to represent a platform in each integration. View [accounting](https://docs.codat.io/integrations/accounting/overview#platform-keys), [banking](https://docs.codat.io/integrations/banking/overview#platform-keys), and [commerce](https://docs.codat.io/integrations/commerce/overview#platform-keys) platform keys.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "platformKey" })