+++ platform/src/sdk/sdk.ts
@@ -33,6 +33,7 @@ export type SDKProps = {
      * The security details required to authenticate the SDK
      */
     security?: shared.Security | (() => Promise<shared.Security>);
+
     /**
      * Allows overriding the default axios client used by the SDK
      */
@@ -60,8 +61,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "1.1.0";
-    genVersion = "2.116.0";
+    sdkVersion = "1.2.0";
+    genVersion = "2.150.0";
+    userAgent = "speakeasy-sdk/typescript 1.2.0 2.150.0 3.0.0 @codat/platform";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);